requirements.txt
appdirs==1.4.3
boto3==1.4.4
botocore==1.5.47
docutils==0.13.1
future==0.16.0
jmespath==0.9.2
oauthlib==2.0.2
packaging==16.8
pyparsing==2.2.0
python-dateutil==2.6.0
python-twitter==3.2.1
requests==2.14.2
requests-oauthlib==0.8.0
s3transfer==0.1.10
six==1.10.0
